LayoutTests/imported/w3c/web-platform-tests/css/css-images/object-view-box-transition-mutation-expected.txt


FAIL Unrelated mutation does not affect object-view-box transition assert_equals: before mutation expected (string) "inset(10px)" but got (undefined) undefined
